* fixed comments
