+++ Home_Work_10/x64/Debug/HW10.tlog/CL.write.1.tlog
