+++ Home_Work_10/x64/Debug/HW10.tlog/Cl.items.tlog
