+++ Home_Work_10/x64/Debug/HW10.tlog/HW10.lastbuildstate
