+++ Home_Work_10/x64/Debug/HW10.tlog/link.read.1.tlog
