+++ Home_Work_10/x64/Debug/HW10.tlog/link.write.1.tlog
